requirements.txt
absl-py==2.1.0
aenum==3.1.15
ai-api-client-sdk==2.1.2
ai-core-sdk==2.3.5
aiofiles==23.2.1
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.28.0
azure-storage-blob==12.17.0
blinker==1.6.2
cachetools==5.3.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
contourpy==1.2.1
cryptography==41.0.2
cycler==0.12.1
dataclasses-json==0.5.12
distro==1.9.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.110.2
ffmpy==0.3.2
filelock==3.13.1
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.51.0
frozenlist==1.4.0
fsspec==2024.2.0
gast==0.5.4
generative-ai-hub-sdk==1.8.0
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.5.2
google-pasta==0.2.0
googleapis-common-protos==1.63.0
gradio==4.37.2
gradio_client==1.0.2
gradio_multimodalchatbot==0.0.1
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
h5py==3.10.0
hdbcli==2.21.26
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
huggingface-hub==0.22.2
idna==3.4
importlib-metadata==7.0.1
importlib_resources==6.4.0
inquirerpy==0.3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.15.0
kiwisolver==1.4.5
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-experimental==0.0.53
langchain-google-genai==1.0.3
langchain-mistralai==0.0.5
langchain-openai==0.1.6
langchain-text-splitters==0.0.2
langchainhub==0.1.20
langsmith==0.1.81
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
matplotlib==3.8.4
mdurl==0.1.2
mistralai==0.1.3
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numexpr==2.8.4
numpy==1.25.1
oauthlib==3.2.2
openai==1.35.9
openapi-schema-pydantic==1.2.4
opt-einsum==3.3.0
orjson==3.9.15
overloading==0.5.0
packaging==23.2
pandas==2.2.1
pfzy==0.3.4
pillow==10.3.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pydot==2.0.0
pydub==0.25.1
Pygments==2.17.2
pyhumps==3.8.0
PyJWT==2.7.0
pyparsing==3.1.1
pypdf==3.13.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.9
pytz==2023.3
PyYAML==6.0.1
referencing==0.34.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruff==0.4.1
safetensors==0.4.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
sqlalchemy-hana==2.2.0
starlette==0.37.2
style==1.1.0
sympy==1.12
tabulate==0.9.0
termcolor==2.4.0
tiktoken==0.5.2
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
#torch==2.2.1
tqdm==4.65.0
#twilio==8.5.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2023.3
update==0.0.1
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.29.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.14.1
yarl==1.9.2
zipp==3.17.0